requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.1
fastapi==0.115.12
frozendict==2.4.6
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
joblib==1.5.1
multidict==6.4.4
multitasking==0.0.11
numpy==2.2.6
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
ta==0.11.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vaderSentiment==3.3.2
websockets==15.0.1
yarl==1.20.0
yfinance==0.2.61
